requirements.txt
ultralytics>=8.2.0
flask>=3.0.0
opencv-python>=4.9.0
cloudinary>=1.39.1
firebase-admin>=6.5.0
python-dotenv>=1.0.1
pillow>=10.3.0
numpy>=1.26.0
requests>=2.31.0
tzdata>=2024.1
